+++ rotkehlchen/data_import/importers/uphold_transactions.py
@@ -1,12 +1,11 @@
 import csv
-import logging
 from pathlib import Path
-from typing import Any
+from typing import TYPE_CHECKING, Any
 
 from rotkehlchen.accounting.structures.balance import Balance
 from rotkehlchen.assets.converters import asset_from_uphold
 from rotkehlchen.constants import ZERO
-from rotkehlchen.data_import.utils import BaseExchangeImporter, hash_csv_row
+from rotkehlchen.data_import.utils import BaseExchangeImporter, SkippedCSVEntry, hash_csv_row
 from rotkehlchen.db.drivers.gevent import DBCursor
 from rotkehlchen.errors.asset import UnknownAsset
 from rotkehlchen.errors.misc import InputError
@@ -15,7 +14,6 @@ from rotkehlchen.exchanges.data_structures import AssetMovement, Trade
 from rotkehlchen.fval import FVal
 from rotkehlchen.history.events.structures.base import HistoryEvent
 from rotkehlchen.history.events.structures.types import HistoryEventSubType, HistoryEventType
-from rotkehlchen.logging import RotkehlchenLogsAdapter
 from rotkehlchen.serialization.deserialize import (
     deserialize_asset_amount,
     deserialize_fee,
@@ -24,14 +22,19 @@ from rotkehlchen.serialization.deserialize import (
 from rotkehlchen.types import AssetAmount, AssetMovementCategory, Fee, Location, Price, TradeType
 from rotkehlchen.utils.misc import ts_sec_to_ms
 
-logger = logging.getLogger(__name__)
-log = RotkehlchenLogsAdapter(logger)
+if TYPE_CHECKING:
+    from rotkehlchen.db.dbhandler import DBHandler
 
 
 UPHOLD_PREFIX = 'UPH_'
 
 
 class UpholdTransactionsImporter(BaseExchangeImporter):
+    """Uphold CSV importer"""
+
+    def __init__(self, db: 'DBHandler') -> None:
+        super().__init__(db=db, name='Uphold')
+
     def _consume_uphold_transaction(
             self,
             write_cursor: DBCursor,
@@ -76,8 +79,7 @@ Activity from uphold with uphold transaction id:
                 elif transaction_type == 'out':
                     event_type = HistoryEventType.SPEND
                 else:
-                    log.debug(f'Ignoring uncaught transaction type of {transaction_type}.')
-                    return
+                    raise SkippedCSVEntry(f'Uncaught transaction type: {transaction_type}.')
                 event = HistoryEvent(
                     event_identifier=f'{UPHOLD_PREFIX}{hash_csv_row(csv_row)}',
                     sequence_index=0,
@@ -110,7 +112,7 @@ Activity from uphold with uphold transaction id:
                     )
                     self.add_trade(write_cursor, trade)
                 else:
-                    log.debug(f'Ignoring trade with Destination Amount: {destination_amount}.')
+                    raise SkippedCSVEntry(f'Trade destination amount is {destination_amount}.')
         elif origin == 'uphold' and transaction_type == 'out':
             if origin_asset == destination_asset:  # Withdrawals
                 asset_movement = AssetMovement(
@@ -142,7 +144,7 @@ Activity from uphold with uphold transaction id:
                 )
                 self.add_trade(write_cursor, trade)
             else:
-                log.debug(f'Ignoring trade with Origin Amount: {origin_amount}.')
+                raise SkippedCSVEntry(f'Trade origin amount is {origin_amount}.')
 
         elif destination == 'uphold' and transaction_type == 'in':
             if origin_asset == destination_asset:  # Deposits
@@ -175,28 +177,38 @@ Activity from uphold with uphold transaction id:
                 )
                 self.add_trade(write_cursor, trade)
             else:
-                log.debug(f'Ignoring trade with Destination Amount: {destination_amount}.')
+                raise SkippedCSVEntry(f'Trade destination amount is {destination_amount}.')
 
     def _import_csv(self, write_cursor: DBCursor, filepath: Path, **kwargs: Any) -> None:
         """
         Information for the values that the columns can have has been obtained from sample CSVs
         """
         with open(filepath, encoding='utf-8-sig') as csvfile:
-            data = csv.DictReader(csvfile)
-            for row in data:
+            for index, row in enumerate(csv.DictReader(csvfile), start=1):
                 try:
+                    self.total_entries += 1
                     self._consume_uphold_transaction(write_cursor, row, **kwargs)
+                    self.imported_entries += 1
                 except UnknownAsset as e:
-                    self.db.msg_aggregator.add_warning(
-                        f'During uphold CSV import found action with unknown '
-                        f'asset {e.identifier}. Ignoring entry',
+                    self.send_message(
+                        row_index=index,
+                        csv_row=row,
+                        msg=f'Unknown asset {e.identifier}.',
+                        is_error=True,
                     )
-                    continue
                 except DeserializationError as e:
-                    self.db.msg_aggregator.add_warning(
-                        f'Deserialization error during uphold CSV import. '
-                        f'{e!s}. Ignoring entry',
+                    self.send_message(
+                        row_index=index,
+                        csv_row=row,
+                        msg=f'Deserialization error: {e!s}.',
+                        is_error=True,
                     )
-                    continue
+                except SkippedCSVEntry as e:
+                    self.send_message(
+                        row_index=index,
+                        csv_row=row,
+                        msg=str(e),
+                        is_error=False,
+                    )
                 except KeyError as e:
                     raise InputError(f'Could not find key {e!s} in csv row {row!s}') from e